Hotfix of setCommandAndCheckForExecutables.
Method setCommandAndCheckForExecutables was stubbed in tests and was never
executed, although in real world it is executed every time someone uses hooks.
Since for different hook languages different parts of the method are summoned,
it was easy for many if/else branches to slip through error/trials or tests.

Due to this it was possible that it contained a lot of typos and errors, such as
callbacks being never called or missing 'this'.