Add failing test for blueprint with glob pattern
See #504

I made sure that blueprints and hooks are working properly when used alone.

There is no problem when using glob pattern for the hook files as long as there is only one blueprint file. Evene if the blueprint file is matched with a wildcard. The problem comes only when there is more than one blueprint file in the list.
On the first place, I found out that error while using a configuration file with dredd. I guess it's the same issue.