feat: Use 127.0.0.1 instead of localhost
Using 'localhost' can provably cause problems on Windows in some cases.
Replacing 'localhost' with '127.0.0.1' avoids these problems. The change
is made everywhere to be consistent and, especially in docs, to present
just one way to do things to Dredd's users. This is a temporary solution
and should be replaced by something, which allows user to specify string
'localhost' as a host under any conditions.

Related also to https://github.com/apiaryio/dredd/issues/721.