feat: Parse hook handler command string, lookup only the command part
If hook handler command string is 'node hello.js --option=1', we want
to find out only whether the 'node' executable is present in the system.
For that reason, we first parse the command and lookup (using 'which')
only the actual command part, without parameters. Looking up the whole
string may or may not work on UNIX/Linux, but it does not work at all
on Windows.