test: Do not construct RegExp with flags from another RegExp
The ps library constructs RegExp with flags (case-insensitive flag,
to be specific) from given search query when looking up system
processes. Constructing RegExp with flags from another RegExp
results in an error in Node 4 (and maybe also other runtimes).
Passing the regular expression to the ps library as a string
solves the problem.