chore: Remove hooks handlers testing
The integration tests of hook handlers were disabled in the past
already due to various issues (#644, #647, #672, #702, #744).
Now there's a security issue in one of the libraries the code
of the tests uses: https://nodesecurity.io/advisories/sync-exec_tmp-files-readable-by-other-users
Dredd already dropped support for Node 0.10 and 0.12, so native
child_process.execSync() could be used instead of the sync-exec
library, but since the code is essentially dead code now, it's
like perform fixes in commented-out code. I decided it would be
the best to remove the code completely for now and to fix it once
someone is re-introducing the test suite.