feat: Upgrade chai from 3.x to 4.x
Bundled chai version starts to be outdated.

BREAKING CHANGE: The chai library is available in the sandboxed JS hooks' interface. For that reason, it is also a part of Dredd's public interface (and for the same reason it's not a devDependency, but dependency). The new version of chai is not backwards compatible in some cases, which means this is a breaking change.