fix: Use plain structure for param values everywhere
Previously, for some legacy reasons, the code for compiling parameters stored
the array of possible values in a structure like this: [{'value': 123}, ...]
@kylef did not know that and did not notice this when rewriting the code
to use the minim library internally, so his changes worked with plain arrays
of values instead: [123, ...] This broke the subsequent code. Interestingly,
neither Dredd Transaction tests or Dredd tests caught this mistake. Especially
Dredd Transactions integration tests were not checking for extraneous warnings
and errors, so the problem slipped through unnoticed. This commit changes the
internal structure for storing the array of values to a plain array.

https://github.com/apiaryio/dredd-transactions/pull/94