test: Do not use array URI parameters in tests
URI parameters in form of arrays are not officially supported in the
API Blueprint parser yet. Even when using Swagger, I found problems
with the current implementation, which prevent Dredd from correctly
expanding the URI Templates. The feature is completely untested on
Dredd side. For that reason, I re-opened the relevant issue on
GitHub and hereby I'm demoting array URI parameters to an experimental
feature, which is not officially fully supported yet.

https://github.com/apiaryio/dredd/issues/791
https://github.com/apiaryio/drafter/issues/500
https://github.com/grncdr/uri-template/issues/17