chore: change the hiring link
The destination of bit.ly short links cannot be changed once created, it is an 'enterprise feature'. I thought this is possible and it's the main reason I used the shortener. The job is expiring after time, so the link will need to change over time again. I'm removing bit.ly for now, fixing the link with this commit, and let's see whether there's better solution in the future.