chore(docs): hardcode the version to 'latest'
We're getting failed docs builds recently, because the npm in RTD's
image isn't able to install npm dependencies successfully - see
https://readthedocs.org/projects/dredd/builds/8343339/ To my surprise,
the latest RTD image contains node@4.2.6 and npm@3.5.2. This means not
only that it might be a bug in the npm, which is could be already
fixed in newer versions, but also that the obvious quick solutions,
such as 'npm ci' or using 'npx', are not available. This brought me
to revisiting why we install npm dependencies in the first place. In the
past the docs had node extensions, but now the only thing needing
the dependencies was the version number, calling semantic-release.
I decided to hardcode the version to 'latest' to avoid the issues
and to remove the complexity.