refactor: remove sandboxed JavaScript hooks
Dredd's sandboxed JS hooks are designed, as opposed to standard JS hooks, to allow running arbitrary code. They provide limited feature set and they take security measures to be able to run the arbitrary code in hooks as a service. I believe the main reason they were introduced to Dredd was the Post-Deployment feature in Apiary, where Dredd was meant to be ran as a service. The feature never made it to general availability though. We had some further ideas about running Dredd as a service, but running hooks as a service was not part of them, at least in the initial phases. The code behind sandboxed hooks in Dredd is rusted and makes development and understanding of Dredd's codebase harder. It uses unmaintained technologies (pitboss). It is also harder to make the JS hooks great again with sandboxed hooks in place. Basically if we ever follow up on the Post-Deployment feature or want to fix issues in Dredd's JS hooks, we need to rewrite everything from the ground up anyway. As the only use case for sandboxed hooks I can imagine is running Dredd as a service, and as even Apiary is not taking advantage of this use case for years, I'm removing the sandboxed hooks from Dredd's codebase.

Close https://github.com/apiaryio/dredd/issues/1178

BREAKING CHANGE: Dredd doesn't support sandboxed JS hooks anymore. Use standard JS hooks instead.