refactor: remove deprecated legacy JS hooks
The legacy interface of JS hooks has been deprecated for ages. It is time for it to finally go.

BREAKING CHANGE: If you have been using the deprecated interface of the JS hooks, you most likely already noticed it years ago, thanks to the warning which Dredd printed out for you. To migrate, use the 'beforeAll' and 'afterAll' hooks with '(transactions, callback)' signature for asynchronous functions, and '(transactions)' signature for synchronous functions.