chore: upgrade chai
This is not a feature, because since we removed the sandboxed JS hooks from the codebase, chai is no more a part of Dredd's public JS API. It still needs to be a dependency though, because it's used in the Transaction Runner. Removing that is up to future refactoring of the JS hooks.