refactor: rename 'configuraton.data' to 'configuration.apiDescriptions'
* feat: add 'apiDescriptionMediaType' to the transaction object

Add 'apiDescriptionMediaType' to the transaction object, turn '.data' into '.apiDescriptions', update docs.

* test: put back incorrectly removed tests

* test: don't use locale for sorting

* refactor: remove 'apiDescriptionMediaType' from the public interface

Because Dredd takes whatever Fury returns as the first media type from the .detect() function, it always gets application/swagger+json for OpenAPI 2, even for YAML files. It's not an issue for Dredd itself, but it is an issue once it's propagated to the transaction object.