docs: add extension for detecting broken links to Dredd's repository
It can happen that a file, which is linked by an absolute GitHub link
from the docs, is removed from a development branch. Such link passes
as not broken with the existing link check, but when a PR with such
branch gets merged into master, the file disappears and the link gets
broken (and master build gets broken). This commit introduces a simple
extension to prevent that.